refactor: voltando os metodos genericos para a sua model
