+++ app/Models/Payment.php
@@ -6,11 +6,12 @@ use App\Http\Traits\SyncPayments;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
+use Illuminate\Support\Arr;
 
 
 class Payment extends Model
 {
-    use HasFactory, SyncPayments;
+    use HasFactory;
 
     protected $fillable = [
         'value',
@@ -22,5 +23,25 @@ class Payment extends Model
     public function payable(): MorphTo
     {
         return $this->morphTo();
+    }
+
+     public static function syncPayments(Model $model, array $paymentsData): void
+    {
+        $incomingIds = array_filter(Arr::pluck($paymentsData, 'id'));
+
+        if (count($incomingIds) > 0) {
+            $model->payments()->whereNotIn('id', $incomingIds)->delete();
+        } else {
+            $model->payments()->delete();
+        }
+
+        foreach ($paymentsData as $paymentData) {
+            $model->payments()->updateOrCreate(
+                [
+                    'id' => $paymentData['id'] ?? null
+                ],
+                $paymentData
+            );
+        }
     }
 }
